requirements.txt
Flask>=2.3.0
google-generativeai>=0.8.0